Move usage function into a separate package and export it
